+++ src/collective/contact/core/indexers.py
@@ -2,6 +2,7 @@ from plone.indexer import indexer
 
 from collective.contact.core.content.held_position import IHeldPosition
 from collective.contact.core.content.organization import IOrganization
+from collective.contact.core.content.position import IPosition
 
 
 @indexer(IOrganization)
@@ -19,3 +20,8 @@ def held_position_searchable_text(obj):
     organization = obj.get_organization()
     indexed_fields.extend(organization.get_organizations_titles())
     return u' '.join(indexed_fields)
+
+
+@indexer(IPosition)
+def position_searchable_text(obj):
+    return obj.SearchableText() + obj.get_organization().Title()