+++ src/collective/contact/core/browser/person.py
@@ -4,6 +4,7 @@ from collective.contact.core.browser.contactable import BaseView
 from collective.contact.core.browser.utils import date_to_DateTime,\
                                                      get_ttw_fields
 from collective.contact.core.interfaces import IContactable
+from collective.contact.core.behaviors import IBirthday
 
 
 class Person(BaseView):
@@ -21,10 +22,13 @@ class Person(BaseView):
         person = self.person
 
         self.name = person.Title()
-        birthday = person.birthday
-        if birthday is not None:
-            birthday = date_to_DateTime(birthday)
-            self.birthday = self.context.toLocalizedTime(birthday)
+        if IBirthday.providedBy(person):
+            birthday = person.birthday
+            if birthday is not None:
+                birthday = date_to_DateTime(birthday)
+                self.birthday = self.context.toLocalizedTime(birthday)
+        else:
+            self.birthday = ""
 
         self.person_title = person.person_title
         self.gender = person.gender or ''