+++ src/collective/contact/core/content/person.py
@@ -3,7 +3,6 @@ from zope.interface import implements
 from zope.interface import Attribute
 from z3c.form.interfaces import NO_VALUE
 from z3c.form.browser.radio import RadioFieldWidget
-from z3c.form.widget import FieldWidget
 
 from five import grok
 
@@ -12,19 +11,11 @@ from plone.dexterity.content import Container
 from plone.dexterity.schema import DexteritySchemaPolicy
 from plone.namedfile.field import NamedImage
 from plone.supermodel import model
-from plone.formwidget.datetime.z3cform import DateWidget
 
 from collective.contact.core import _
 from collective.contact.core.browser.contactable import Contactable
 from collective.contact.widget.interfaces import IContactContent
 from collective.contact.core.content.held_position import IHeldPosition
-
-
-def DateFieldWidget(field, request):
-    """IFieldWidget factory for DatetimeWidget."""
-    widget = FieldWidget(field, DateWidget(request))
-    widget.years_range = (-200, 1)
-    return widget
 
 
 class IPerson(model.Schema, IContactContent):
@@ -50,11 +41,6 @@ class IPerson(model.Schema, IContactContent):
         title=_("Person title"),
         required=False,
         )
-    form.widget(birthday=DateFieldWidget)
-    birthday = schema.Date(
-        title=_("Birthday"),
-        required=False,
-        )
     photo = NamedImage(
         title=_("Photo"),
         required=False,
@@ -63,6 +49,7 @@ class IPerson(model.Schema, IContactContent):
     def get_held_positions(self):
         """Returns held positions of this person
         """
+
 
 class PersonContactableAdapter(Contactable):
     """Contactable adapter for Person content type"""
@@ -97,7 +84,10 @@ class Person(Container):
 
     def get_held_positions(self):
         return [obj for obj in self.values() if IHeldPosition.providedBy(obj)]
-
+    
+    def get_held_positions_titles(self):
+        return [p.Title() for p in self.get_held_positions()]
+    
 
 class PersonSchemaPolicy(grok.GlobalUtility,
                          DexteritySchemaPolicy):