+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 1.0 (unreleased)
 ----------------
 
+- Display position label in title of held position view page.
+  [thomasdesvenain]
+
 - Added an additional input text label to held positions,
   used on titles if held_position is directly related to an organization.
   [thomasdesvenain]